clean: move Footnotes struct to other file
