+++ markdown/src/translate.rs
@@ -1,5 +1,6 @@
 pub mod element;
 pub mod error;
+pub mod footnotes;
 pub mod node;
 pub mod translator;
